feat: split EXESS implementation into module-specific files
Move the EXESS geo-opt and QM/MM wrappers into dedicated top-level
modules, keep the energy-facing helpers in rush.exess, and update the
surrounding docs/examples/tests to use the new import structure.

This also cleans up the EXESS tutorial and reference snippets so they
match the current output shapes and helper APIs.
