+++ docs/guides/02-first-run.md
@@ -11,16 +11,10 @@ Each Rush module has a Python submodule that provides support for it. For exampl
 from rush import exess
 ```
 
-You can then access the function that directly wraps EXESS as expected:
+For example, single-point energy calculations can be performred using `exess_energy`:
 ```python
-from rush import exess
-exess.exess("input_topology.json", collect=True)
-```
-
-Or the possibly more clear:
-```python
-from rush.exess import exess as run_exess
-run_exess("input_topology.json", collect=True)
+from rush.exess import exess_energy
+exess_energy("input_topology.json", collect=True)
 ```
 
 And you're done! The EXESS module has a rich set of capabilities and configuration, but by default will perform an energy calculation on the input topology molecule. To learn more about EXESS capabilities, see the [EXESS documentation](../exess/index.md). To learn about TRC files, check the [topologies section](../exess/topologies.md).
@@ -34,9 +28,10 @@ Here's a complete example that runs an energy calculation and prints the result:
 ```python
 import json
 from rush import exess
+from rush.exess import exess_energy
 from rush.client import save_object
 
-res = exess.energy("benzene_t.json", collect=True)
+res = exess_energy("benzene_t.json", collect=True)
 output_file = save_object(res[0]["path"])
 
 with open(output_file) as f:
@@ -51,8 +46,8 @@ print(output_data["qmmbe"]["expanded_hf_energy"])
 Rush modules can take a long time, so by default run asynchronously: the function that triggers the run will return once the run is submitted. In order to obtain the output synchronously for this same call, pass `collect=True` as we've done above. You can also collect the run later:
 ```python
 from rush.client import collect_run
-from rush.exess import exess as run_exess
-id = run_exess("input_topology.json")
+from rush.exess import exess_energy
+id = exess_energy("input_topology.json")
 result = collect_run(id)
 ```
 
@@ -63,11 +58,13 @@ As shown, module calls return a "run ID" that can be used as above to collect an
 
 Or, call one of the other supplementary EXESS entry points, built to facilitate using EXESS's various capabilities:
 ```python
-from rush import exess
-exess.energy(...)
-exess.interaction_energy(...)
-exess.optimization(...)
-exess.qmmm(...)
+from rush.exess import exess_energy, exess_interaction_energy
+from rush.exess_geo_opt import exess_geo_opt
+from rush.exess_qmmm import exess_qmmm
+exess_energy(...)
+exess_interaction_energy(...)
+exess_geo_opt(...)
+exess_qmmm(...)
 ```
 
 ## View your run in the Rush Web Interface
