+++ docs/modules.rst
@@ -4,6 +4,8 @@ Rush Modules
    :maxdepth: 1
 
    rush.exess
+   rush.exess_geo_opt
+   rush.exess_qmmm
    rush.nnxtb
    rush.prepare_protein
    rush.prepare_complex
