+++ docs/rush.exess_geo_opt.rst
@@ -1,0 +1,8 @@
+EXESS Geometry Optimization
+===========================
+
+.. automodule:: rush.exess_geo_opt
+   :members:
+   :member-order: bysource
+   :show-inheritance:
+   :undoc-members:
