+++ docs/rush.exess_qmmm.rst
@@ -1,0 +1,8 @@
+EXESS QMMM
+==========
+
+.. automodule:: rush.exess_qmmm
+   :members:
+   :member-order: bysource
+   :show-inheritance:
+   :undoc-members:
