+++ src/rush/exess.py
@@ -3,19 +3,21 @@
 EXESS module helpers for the Rush Python client.
 
 EXESS supports whole-system energy calculations (fragmented or unfragmented),
-interaction energy between a fragment and the rest of the system, geometry
-optimization, simulations, and gradient/Hessian calculations. It supports
-multiple levels of theory (e.g., restricted/unrestricted HF, RI-MP2, DFT),
-flexible basis set selection, and configurable n-mer fragmentation levels.
+interaction energy between a fragment and the rest of the system, and
+gradient/Hessian calculations. It supports multiple levels of theory
+(e.g., restricted/unrestricted HF, RI-MP2, DFT), flexible basis set
+selection, and configurable n-mer fragmentation levels.
 
 Quick Links
 -----------
 
 - :func:`rush.exess.exess`
-- :func:`rush.exess.energy`
-- :func:`rush.exess.interaction_energy`
-- :func:`rush.exess.qmmm`
-- :func:`rush.exess.optimization`
+- :func:`rush.exess.exess_energy`
+- :func:`rush.exess.exess_interaction_energy`
+- :func:`rush.exess.fetch_outputs`
+- :func:`rush.exess.save_outputs`
+- :mod:`rush.exess_geo_opt`
+- :mod:`rush.exess_qmmm`
 """
 
 import enum
@@ -1012,83 +1014,9 @@ class KSDFTKeywords:
         )
 
 
-@dataclass
-class Trajectory:
-    """
-    Configure the output of QMMM runs. By default, will provide all atoms at every frame.
-    """
-
-    #: Save every n frames to the trajectory, where n is the interval specified.
-    interval: int | None = None
-    #: The frame at which to start the trajectory.
-    start: int | None = None
-    #: The frame at which to end the trajectory.
-    end: int | None = None
-    #: Whether to include waters in the trajectory. Convenient for reducing output size.
-    include_waters: int | None = None
-
-    def _to_rex(self):
-        return Template(
-            """Some (exess_qmmm_rex::MDTrajectory {
-              format = None,
-              interval = $maybe_interval,
-              start = $maybe_start,
-              end = $maybe_end,
-              include_waters = $maybe_include_waters,
-            })"""
-        ).substitute(
-            maybe_interval=optional_str(self.interval),
-            maybe_start=optional_str(self.start),
-            maybe_end=optional_str(self.end),
-            maybe_include_waters=optional_str(self.include_waters),
-        )
-
-
-@dataclass
-class Restraints:
-    """
-    Restrain atoms using an external force proportional to its distance from its original position,
-    scaled by `k` (larger values mean a stronger restraint).
-
-    All atoms can be fixed by specifying `free_atoms = []`.
-    """
-
-    #: Scaling factor for restraints (larger values mean a stronger restraint).
-    k: float | None = None
-    #: Which atoms to hold fixed. All fixed/free parameters are mutually exclusive.
-    fixed_atoms: list[int] | None = None
-    #: Which atoms to keep unfixed. All fixed/free parameters are mutually exclusive.
-    free_atoms: list[int] | None = None
-    #: Which fragments to hold fixed. All fixed/free parameters are mutually exclusive.
-    fixed_fragments: list[int] | None = None
-    #: Which fragments to keep unfixed. All fixed/free parameters are mutually exclusive.
-    free_fragments: list[int] | None = None
-    #: Flag to easily enable fixing all heavy atoms only. Mutually exclusive with fixed/free parameters.
-    fix_heavy: bool | None = None
-
-    def _to_rex(self):
-        return Template(
-            """Some (exess_rex::Restraints {
-              k = $maybe_k,
-              fixed_atoms = $maybe_fixed_atoms,
-              free_atoms = $maybe_free_atoms,
-              fixed_fragments = $maybe_fixed_fragments,
-              free_fragments = $maybe_free_fragments,
-              fix_heavy = $maybe_fix_heavy,
-            })"""
-        ).substitute(
-            maybe_k=optional_str(self.k),
-            maybe_fixed_atoms=optional_str(self.fixed_atoms),
-            maybe_free_atoms=optional_str(self.free_atoms),
-            maybe_fixed_fragments=optional_str(self.fixed_fragments),
-            maybe_free_fragments=optional_str(self.free_fragments),
-            maybe_fix_heavy=optional_str(self.fix_heavy),
-        )
-
-
 def exess(
     topology_path: Path | str,
-    driver: str = "Energy",
+    driver: str,
     method: MethodT = "RestrictedKSDFT",
     basis: BasisT = "cc-pVDZ",
     aux_basis: AuxBasisT | None = None,
@@ -1196,7 +1124,7 @@ in
                 print(f"Error: {error['message']}", file=sys.stderr)
 
 
-def energy(
+def exess_energy(
     topology_path: Path | str,
     method: MethodT = "RestrictedKSDFT",
     basis: BasisT = "cc-pVDZ",
@@ -1243,7 +1171,7 @@ def _split_result_outputs(
         return (res, None)
 
     if len(res) == 0 or len(res) > 2:
-        raise ValueError("exess.energy should return 1 or 2 outputs.")
+        raise ValueError("exess_energy should return 1 or 2 outputs.")
 
     return (res[0], res[1] if len(res) == 2 else None)
 
@@ -1318,7 +1246,7 @@ def save_outputs(
     """
     Download and save energy calculation outputs from Rush.
 
-    Takes the result from exess functions (energy(), interaction_energy(), etc.)
+    Takes the result from exess functions (`exess_energy()`, `exess_interaction_energy()`, etc.)
     and downloads the output files to disk. The result can be either a tuple or list
     because collect_run() returns a list when there are multiple outputs, which is
     converted to a tuple for consistent processing.
@@ -1380,7 +1308,7 @@ def save_energy_outputs(
     return save_outputs(res, extract=extract)
 
 
-def interaction_energy(
+def exess_interaction_energy(
     topology_path: Path | str,
     reference_fragment: int,
     method: MethodT = "RestrictedKSDFT",
@@ -1400,7 +1328,7 @@ def interaction_energy(
     Compute the interaction energy between the fragment with index `reference_fragment` and the rest of the system
     in the toplogy file at `topology_path`.
     """
-    return energy(
+    return exess_energy(
         topology_path=topology_path,
         method=method,
         basis=basis,
@@ -1418,501 +1346,3 @@ def interaction_energy(
         run_opts=run_opts,
         collect=collect,
     )
-
-
-def qmmm(
-    topology_path: Path | str,
-    n_timesteps: int,
-    residues_path: Path | str | None = None,
-    dt_ps: float = 2e-3,
-    temperature_kelvin: float = 290.0,
-    pressure_atm: float | None = None,
-    restraints: Restraints | None = None,
-    trajectory: Trajectory = Trajectory(),
-    gradient_finite_difference_step_size: float | None = None,
-    method: MethodT = "RestrictedKSDFT",
-    basis: BasisT = "cc-pVDZ",
-    aux_basis: AuxBasisT | None = None,
-    standard_orientation: StandardOrientationT | None = None,
-    force_cartesian_basis_sets: bool | None = None,
-    scf_keywords: SCFKeywords | None = None,
-    frag_keywords: FragKeywords = FragKeywords(),
-    ksdft_keywords: KSDFTKeywords | _KSDFTDefault | None = _KSDFTDefault.DEFAULT,
-    qm_fragments: list[int] | None = None,
-    mm_fragments: list[int] | None = None,
-    system: System | None = None,
-    run_spec: RunSpec = RunSpec(gpus=1),
-    run_opts: RunOpts = RunOpts(),
-    collect: bool = False,
-):
-    """
-    Run a QMMM simulation of the system in the QDX topology and residues files at `topology_path` and `residues_path`.
-
-    Specifying the number of timesteps is mandatory.
-    If pressure is None, an NVT ensemble is used; if pressure is specified, an NPT ensemble is used.
-    Fragments can be specified as QM or MM fragments via the respective parameters.
-    If one fragment list parameter is specified, the rest of the fragments are inferred to be of the other type.
-    If both fragment list parameters are specified, each fragment must be placed in exactly one of the lists.
-    """
-    ksdft_keywords = KSDFTKeywords.resolve(ksdft_keywords, method)
-
-    # Upload inputs
-    topology_vobj = upload_object(topology_path)
-    residues_vobj = None
-    if residues_path is not None:
-        residues_vobj = upload_object(residues_path)
-
-    # Run rex
-    rex = Template("""let
-  obj_j = λ j →
-    VirtualObject { path = j, format = ObjectFormat::json, size = 0 },
-  exess = λ topology residues →
-    exess_qmmm_rex_s
-      ($run_spec)
-      (exess_qmmm_rex::QMMMParams {
-        schema_version = "0.2.0",
-        model = Some (exess_qmmm_rex::Model {
-          method = exess_qmmm_rex::Method::$method,
-          basis = "$basis",
-          aux_basis = $maybe_aux_basis,
-          standard_orientation = $maybe_standard_orientation,
-          force_cartesian_basis_sets = $maybe_force_cartesian_basis_sets,
-        }),
-        system = $system,
-        keywords = exess_qmmm_rex::Keywords {
-          scf = $maybe_scf_keywords,
-          ks_dft = $maybe_ks_keywords,
-          rtat = None,
-          frag = $maybe_frag_keywords,
-          boundary = None,
-          log = None,
-          dynamics = None,
-          integrals = None,
-          debug = None,
-          export = None,
-          guess = None,
-          force_field = None,
-          optimization = None,
-          hessian = None,
-          gradient = Some (exess_qmmm_rex::GradientKeywords {
-            finite_difference_step_size = $maybe_gradient_finite_difference_step_size,
-            method = Some exess_qmmm_rex::DerivativesMethod::Analytical,
-          }),
-          qmmm = Some (exess_qmmm_rex::QMMMKeywords {
-            n_timesteps = $n_timesteps,
-            dt_ps = $dt_ps,
-            temperature_kelvin = $temperature_kelvin,
-            pressure_atm = $maybe_pressure_atm,
-            minimisation = None,
-            trajectory = $trajectory,
-            restraints = $maybe_restraints,
-            energy_csv = None,
-          }),
-          machine_learning = None,
-          regions = $maybe_regions,
-        },
-      })
-      (obj_j topology)
-      (Some (obj_j residues))
-in
-  exess "$topology_vobj_path" "$residues_vobj_path"
-""").substitute(
-        run_spec=run_spec._to_rex(),
-        method=method,
-        basis=basis,
-        maybe_aux_basis=optional_str(aux_basis),
-        maybe_standard_orientation=optional_str(
-            standard_orientation, "exess_rex::StandardOrientation::"
-        ),
-        maybe_force_cartesian_basis_sets=optional_str(force_cartesian_basis_sets),
-        system=system._to_rex() if system is not None else "None",
-        maybe_scf_keywords=(
-            scf_keywords._to_rex() if scf_keywords is not None else "None"
-        ),
-        maybe_ks_keywords=(
-            ksdft_keywords._to_rex() if ksdft_keywords is not None else "None"
-        ),
-        maybe_frag_keywords=(
-            frag_keywords._to_rex() if frag_keywords is not None else "None"
-        ),
-        maybe_gradient_finite_difference_step_size=optional_str(
-            gradient_finite_difference_step_size
-        ),
-        n_timesteps=n_timesteps,
-        dt_ps=dt_ps,
-        temperature_kelvin=temperature_kelvin,
-        maybe_pressure_atm=optional_str(pressure_atm),
-        trajectory=trajectory._to_rex(),
-        maybe_restraints=restraints._to_rex() if restraints is not None else "None",
-        maybe_regions=(
-            Template(
-                """Some (exess_qmmm_rex::RegionKeywords {
-            qm_fragments = $maybe_qm_fragments,
-            mm_fragments = $maybe_mm_fragments,
-            ml_fragments = Some [],
-          })"""
-            ).substitute(
-                maybe_qm_fragments=optional_str(qm_fragments),
-                maybe_mm_fragments=optional_str(mm_fragments),
-            )
-            if not (qm_fragments is None and mm_fragments is None)
-            else "None"
-        ),
-        topology_vobj_path=topology_vobj["path"],
-        residues_vobj_path=residues_vobj["path"] if residues_vobj is not None else "",
-    )
-    try:
-        run_id = _submit_rex(_get_project_id(), rex, run_opts)
-        if collect:
-            return collect_run(run_id)
-        else:
-            return run_id
-
-    except TransportQueryError as e:
-        if e.errors:
-            for error in e.errors:
-                print(f"Error: {error['message']}", file=sys.stderr)
-
-
-@dataclass
-class OptimizationConvergenceCriteria:
-    metric: str | None = None
-    gradient_threshold: float | None = None
-    delta_energy_threshold: float | None = None
-    step_component_threshold: float | None = None
-
-    def _to_rex(self, reference_fragment: int | None = None):
-        return Template(
-            """Some (exess_geo_opt_rex::OptimizationConvergenceCriteria {
-            metric = $maybe_metric,
-            gradient_threshold = $maybe_gradient_threshold,
-            delta_energy_threshold = $maybe_delta_energy_threshold,
-            step_component_threshold = $maybe_step_component_threshold,
-          })"""
-        ).substitute(
-            maybe_metric=optional_str(self.metric),  # TODO: enum prefix
-            maybe_gradient_threshold=optional_str(self.gradient_threshold),
-            maybe_delta_energy_threshold=optional_str(self.delta_energy_threshold),
-            maybe_step_component_threshold=optional_str(self.step_component_threshold),
-        )
-
-
-type CoordinateSystemT = Literal["Cartesian", "NaturalInternal", "DelocalisedInternal"]
-
-type HessianGuessTypeT = Literal["Identity", "ScaledIdentity", "Schlegel", "Lindh"]
-
-type OptimizationAlgorithmTypeT = Literal[
-    "EigenvectorFollowing", "TrustRegionAugmentedHessian", "LBFGS"
-]
-
-
-@dataclass
-class TrustRegionKeywords:
-    initial_radius: float | None = None
-    max_radius: float | None = None
-    min_radius: float | None = None
-    increase_factor: float | None = None
-    decrease_factor: float | None = None
-    constrict_factor: float | None = None
-    increase_threshold: float | None = None
-    decrease_threshold: float | None = None
-    rejection_threshold: float | None = None
-
-    def _to_rex(self):
-        return Template(
-            """Some (exess_geo_opt_rex::TrustRegionKeywords {
-            initial_radius = $maybe_initial_radius,
-            max_radius = $maybe_max_radius,
-            min_radius = $maybe_min_radius,
-            increase_factor = $maybe_increase_factor,
-            decrease_factor = $maybe_decrease_factor,
-            constrict_factor = $maybe_constrict_factor,
-            increase_threshold = $maybe_increase_threshold,
-            decrease_threshold = $maybe_decrease_threshold,
-            rejection_threshold = $maybe_rejection_threshold,
-          })"""
-        ).substitute(
-            maybe_initial_radius=optional_str(self.initial_radius),
-            maybe_max_radius=optional_str(self.max_radius),
-            maybe_min_radius=optional_str(self.min_radius),
-            maybe_increase_factor=optional_str(self.increase_factor),
-            maybe_decrease_factor=optional_str(self.decrease_factor),
-            maybe_constrict_factor=optional_str(self.constrict_factor),
-            maybe_increase_threshold=optional_str(self.increase_threshold),
-            maybe_decrease_threshold=optional_str(self.decrease_threshold),
-            maybe_rejection_threshold=optional_str(self.rejection_threshold),
-        )
-
-
-type LBFGSLinesearchT = Literal[
-    "MoreThuente", "BacktrackingArmijo", "BacktrackingWolfe", "BacktrackingStrongWolfe"
-]
-
-
-@dataclass
-class LBFGSKeywords:
-    linesearch: LBFGSLinesearchT = "BacktrackingStrongWolfe"
-    n_corrections: int | None = None
-    epsilon: float | None = None
-    max_linesearch: int | None = None
-    gtol: float | None = None
-
-    def _to_rex(self):
-        return Template(
-            """Some (exess_geo_opt_rex::LBFGSKeywords {
-              linesearch = $maybe_linesearch,
-              n_corrections = $maybe_n_corrections,
-              epsilon = $maybe_epsilon,
-              max_linesearch = $maybe_max_linesearch,
-              gtol = $maybe_gtol,
-            })"""
-        ).substitute(
-            maybe_linesearch=optional_str(
-                self.linesearch, "exess_geo_opt_rex::LBFGSLinesearch::"
-            ),
-            maybe_n_corrections=optional_str(self.n_corrections),
-            maybe_epsilon=optional_str(self.epsilon),
-            maybe_max_linesearch=optional_str(self.max_linesearch),
-            maybe_gtol=optional_str(self.gtol),
-        )
-
-
-@dataclass
-class OptimizationKeywords:
-    convergence_criteria: OptimizationConvergenceCriteria | None = None
-    optimizer_reset_interval: int | None = None
-    coordinate_system: CoordinateSystemT | None = None
-    constraints: list[list[int]] | None = None
-    hessian_guess: HessianGuessTypeT | None = None
-    algorithm: OptimizationAlgorithmTypeT | None = None
-    lbfgs_keywords: LBFGSKeywords | None = None
-    frozen_distance_slippage_tolerance_angstroms: float | None = None
-    frozen_angle_slippage_tolerance_degrees: float | None = None
-    trust_region_keywords: TrustRegionKeywords | None = None
-    fixed_atoms: list[int] | None = None
-    free_atoms: list[int] | None = None
-    fixed_fragments: list[int] | None = None
-    free_fragments: list[int] | None = None
-    fix_heavy: bool | None = None
-
-    def _to_rex(self, max_iters):
-        return Template(
-            """Some (exess_geo_opt_rex::OptimizationKeywords {
-            max_iters = $max_iters,
-            convergence_criteria = $maybe_convergence_criteria,
-            optimizer_reset_interval = $maybe_optimizer_reset_interval,
-            coordinate_system = $maybe_coordinate_system,
-            constraints = $maybe_constraints,
-            hessian_guess = $maybe_hessian_guess,
-            algorithm = $maybe_algorithm,
-            lbfgs_keywords = $maybe_lbfgs_keywords,
-            frozen_distance_slippage_tolerance_angstroms = $maybe_frozen_distance_slippage_tolerance_angstroms,
-            frozen_angle_slippage_tolerance_degrees = $maybe_frozen_angle_slippage_tolerance_degrees,
-            trust_region_keywords = $maybe_trust_region_keywords,
-            fixed_atoms = $maybe_fixed_atoms,
-            free_atoms = $maybe_free_atoms,
-            fixed_fragments = $maybe_fixed_fragments,
-            free_fragments = $maybe_free_fragments,
-            fix_heavy = $maybe_fix_heavy,
-          })"""
-        ).substitute(
-            max_iters=max_iters,
-            maybe_convergence_criteria=(
-                self.convergence_criteria._to_rex()
-                if self.convergence_criteria is not None
-                else "None"
-            ),
-            maybe_optimizer_reset_interval=optional_str(self.optimizer_reset_interval),
-            maybe_coordinate_system=optional_str(
-                self.coordinate_system, "exess_geo_opt_rex::CoordinateSystem::"
-            ),
-            # maybe_constraints=optional_list(
-            #     self.constraints,
-            #     lambda constraint: f"vec![{', '.join(f'exess_geo_opt_rex::AtomRef ({atom})' for atom in constraint)}]",
-            # ),
-            maybe_constraints="None",  # TODO
-            maybe_hessian_guess=optional_str(
-                self.hessian_guess, "exess_geo_opt_rex::HessianGuessType::"
-            ),
-            maybe_algorithm=optional_str(
-                self.algorithm, "exess_geo_opt_rex::OptimizationAlgorithmType::"
-            ),
-            maybe_lbfgs_keywords=(
-                self.lbfgs_keywords._to_rex()
-                if self.lbfgs_keywords is not None
-                else "None"
-            ),
-            maybe_frozen_distance_slippage_tolerance_angstroms=optional_str(
-                self.frozen_distance_slippage_tolerance_angstroms
-            ),
-            maybe_frozen_angle_slippage_tolerance_degrees=optional_str(
-                self.frozen_angle_slippage_tolerance_degrees
-            ),
-            maybe_trust_region_keywords=(
-                self.trust_region_keywords._to_rex()
-                if self.trust_region_keywords is not None
-                else "None"
-            ),
-            maybe_fixed_atoms=optional_str(self.fixed_atoms),
-            maybe_free_atoms=optional_str(self.free_atoms),
-            maybe_fixed_fragments=optional_str(self.fixed_fragments),
-            maybe_free_fragments=optional_str(self.free_fragments),
-            maybe_fix_heavy=optional_str(self.fix_heavy),
-        )
-
-
-def optimization(
-    topology_path: Path | str,
-    max_iters: int,
-    residues_path: Path | str | None = None,
-    optimization_keywords: OptimizationKeywords = OptimizationKeywords(),
-    method: MethodT = "RestrictedKSDFT",
-    basis: BasisT = "cc-pVDZ",
-    aux_basis: AuxBasisT | None = None,
-    standard_orientation: StandardOrientationT | None = None,
-    force_cartesian_basis_sets: bool | None = None,
-    scf_keywords: SCFKeywords | None = None,
-    ksdft_keywords: KSDFTKeywords | _KSDFTDefault | None = _KSDFTDefault.DEFAULT,
-    qm_fragments: list[int] | None = None,
-    mm_fragments: list[int] | None = None,
-    system: System | None = None,
-    run_spec: RunSpec = RunSpec(gpus=1),
-    run_opts: RunOpts = RunOpts(),
-    collect: bool = False,
-):
-    """
-    Run optimization on the system in the QDX topology and residues files at `topology_path`.
-
-    Specifying the maximum iterations is mandatory.
-    Fragment-based QM calculation is not supported, but fragments can be used for specifying regions as QM or MM.
-    If one fragment list parameter is specified, the rest of the fragments are inferred to be of the other type.
-    If both fragment list parameters are specified, each fragment must be placed in exactly one of the lists.
-    """
-    ksdft_keywords = KSDFTKeywords.resolve(ksdft_keywords, method)
-
-    # Upload inputs
-    topology_vobj = upload_object(topology_path)
-    residues_vobj = None
-    if residues_path is not None:
-        residues_vobj = upload_object(residues_path)
-
-    # Run rex
-    rex = Template("""let
-  obj_j = λ j →
-    VirtualObject { path = j, format = ObjectFormat::json, size = 0 },
-  exess = λ topology residues →
-    exess_geo_opt_rex_s
-      ($run_spec)
-      (exess_geo_opt_rex::OptimizationParams {
-        schema_version = "0.2.0",
-        external_charges = None,
-        model = Some (exess_geo_opt_rex::Model {
-          method = exess_geo_opt_rex::Method::$method,
-          basis = "$basis",
-          aux_basis = $maybe_aux_basis,
-          standard_orientation = $maybe_standard_orientation,
-          force_cartesian_basis_sets = $maybe_force_cartesian_basis_sets,
-        }),
-        system = $maybe_system,
-        keywords = exess_geo_opt_rex::Keywords {
-          scf = $maybe_scf_keywords,
-          ks_dft = $maybe_ks_keywords,
-          rtat = None,
-          frag = None,
-          boundary = None,
-          log = None,
-          dynamics = None,
-          integrals = None,
-          debug = None,
-          export = None,
-          guess = None,
-          force_field = None,
-          optimization = $maybe_optimization_keywords,
-          hessian = None,
-          gradient = None,
-          qmmm = $maybe_qmmm_keywords,
-          machine_learning = None,
-          regions = $maybe_regions,
-        },
-      })
-      [ (obj_j topology) ]
-      $residues_expr
-in
-  exess "$topology_vobj_path" "$residues_vobj_path"
-""").substitute(
-        run_spec=run_spec._to_rex(),
-        method=method,
-        basis=basis,
-        maybe_aux_basis=optional_str(aux_basis),
-        maybe_standard_orientation=optional_str(
-            standard_orientation, "exess_rex::StandardOrientation::"
-        ),
-        maybe_force_cartesian_basis_sets=optional_str(force_cartesian_basis_sets),
-        maybe_system=system._to_rex() if system is not None else "None",
-        maybe_scf_keywords=(
-            scf_keywords._to_rex() if scf_keywords is not None else "None"
-        ),
-        maybe_ks_keywords=(
-            ksdft_keywords._to_rex() if ksdft_keywords is not None else "None"
-        ),
-        maybe_optimization_keywords=(
-            optimization_keywords._to_rex(max_iters)
-            if optimization_keywords is not None
-            else "None"
-        ),
-        maybe_qmmm_keywords=(
-            """Some (exess_qmmm_rex::QMMMKeywords {
-            n_timesteps = 1,
-            dt_ps = 0.002,
-            temperature_kelvin = 290.0,
-            pressure_atm = None,
-            minimisation = None,
-            trajectory = None,
-            restraints = None,
-            energy_csv = None,
-          })"""
-            if mm_fragments or (qm_fragments is not None)
-            else "None"
-        ),
-        maybe_regions=(
-            Template(
-                """Some (exess_qmmm_rex::RegionKeywords {
-            qm_fragments = $maybe_qm_fragments,
-            mm_fragments = $maybe_mm_fragments,
-            ml_fragments = Some [],
-          })"""
-            ).substitute(
-                maybe_qm_fragments=optional_str(qm_fragments),
-                maybe_mm_fragments=optional_str(mm_fragments),
-            )
-            if not (qm_fragments is None and mm_fragments is None)
-            else "None"
-        ),
-        residues_expr=(
-            "(Some [ (obj_j residues) ])" if residues_path is not None else "None"
-        ),
-        topology_vobj_path=topology_vobj["path"],
-        residues_vobj_path=residues_vobj["path"] if residues_vobj is not None else "",
-    )
-    try:
-        run_id = _submit_rex(_get_project_id(), rex, run_opts)
-        if collect:
-            return collect_run(run_id)
-        else:
-            return run_id
-
-    except TransportQueryError as e:
-        if e.errors:
-            for error in e.errors:
-                print(f"Error: {error['message']}", file=sys.stderr)
-
-
-# TODO:
-#  - trace for failure
-#  - stdout, stderr
-#  - other module instance info?
-#  - qmmm minimisation config:
-#    minimisation = Some (exess_rex::ClassicalMinimisation {
-#      err_tol_kj_per_mol_nm = $err_tol_kj_per_mol_nm,
-#      max_iterations = $max_iterations,
-#    }),
