+++ tests/conftest.py
@@ -14,6 +14,7 @@ FAST_TEST_PREFIXES = (
     "test_merge",
     "test_fetch_runs",
     "test_client_collect_run",
+    "test_exess_namespaces",
     "test_exess_output_helpers",
     "test_nnxtb_output_helpers",
     "test_auto3d_output_helpers",
