+++ tests/test_exess_chelpg.py
@@ -6,6 +6,7 @@ from pprint import pp
 
 from rush import exess, from_json
 from rush.client import RunOpts, fetch_object, set_opts
+from rush.exess import exess_energy
 
 
 def test_exess_energy_chelpg_1hsg_MK1():
@@ -18,7 +19,7 @@ def test_exess_energy_chelpg_1hsg_MK1():
         json.dump(trc.topology.to_json(), tf)
         topology_path = tf.name
 
-    res = exess.energy(
+    res = exess_energy(
         topology_path,
         basis="PCSeg-0",
         frag_keywords=None,  # Important, to disable fragmentation
@@ -38,7 +39,7 @@ def test_exess_energy_chelpg_1hsg_MK1():
 def test_exess_energy_chelpg_benzene():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path(__file__).parent / "data"
-    res = exess.energy(
+    res = exess_energy(
         data_dir / "benzene_t.json",
         method="RestrictedRIMP2",
         basis="def2-TZVP",
