+++ tests/test_exess_energy_dft.py
@@ -3,12 +3,13 @@ from pathlib import Path
 
 from rush import exess
 from rush.client import RunOpts, collect_run, set_opts
+from rush.exess import exess_energy
 
 
 def test_exess_energy_dft_hyb():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path(__file__).parent / "data"
-    id = exess.energy(
+    id = exess_energy(
         data_dir / "benzene_t.json",
         method="RestrictedKSDFT",
         ksdft_keywords=exess.KSDFTKeywords(
@@ -40,7 +41,7 @@ def test_exess_energy_dft_hyb():
 def test_exess_energy_dft_dhyb():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path(__file__).parent / "data"
-    id = exess.energy(
+    id = exess_energy(
         data_dir / "benzene_t.json",
         method="RestrictedKSDFT",
         basis="cc-pVTZ",
