+++ tests/test_exess_interaction_energy_setonix.py
@@ -5,6 +5,7 @@ import pytest
 
 from rush import Topology, exess
 from rush.client import RunOpts, RunSpec, set_opts
+from rush.exess import exess_interaction_energy
 
 
 @pytest.mark.timeout(1800)
@@ -14,7 +15,7 @@ def test_exess_interaction_energy_setonix():
     topology = Topology.from_json(data_dir / "tyk2_ejm_31_t.json")
     lig_idx = 93
     frag_idcs = topology.get_fragments_near_fragment(lig_idx, 6.0) + [lig_idx]
-    res = exess.interaction_energy(
+    res = exess_interaction_energy(
         data_dir / "tyk2_ejm_31_t.json",
         lig_idx,
         basis="PCSeg-0",
