+++ tests/test_exess_optimization_qm+mm.py
@@ -1,19 +1,20 @@
 import sys
 from pathlib import Path
 
-from rush import exess
+from rush import exess, exess_geo_opt
 from rush.client import RunOpts, save_object, set_opts
+from rush.exess_geo_opt import exess_geo_opt as run_exess_geo_opt
 
 
 def test_exess_optimization_qm_mm():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path.cwd() / "tests" / "data"
-    res = exess.optimization(
+    res = run_exess_geo_opt(
         max_iters=100,
-        optimization_keywords=exess.OptimizationKeywords(
+        optimization_keywords=exess_geo_opt.OptimizationKeywords(
             coordinate_system="Cartesian",
             algorithm="LBFGS",
-            lbfgs_keywords=exess.LBFGSKeywords(),
+            lbfgs_keywords=exess_geo_opt.LBFGSKeywords(),
         ),
         topology_path=data_dir / "6a5j_t.json",
         residues_path=data_dir / "6a5j_r.json",
