+++ tests/test_exess_optimization_qm.py
@@ -1,17 +1,18 @@
 import sys
 from pathlib import Path
 
-from rush import exess
+from rush import exess_geo_opt
 from rush.client import RunOpts, save_object, set_opts
+from rush.exess_geo_opt import exess_geo_opt as run_exess_geo_opt
 
 
 def test_exess_optimization_qm():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path.cwd() / "tests" / "data"
-    res = exess.optimization(
+    res = run_exess_geo_opt(
         max_iters=100,
         topology_path=data_dir / "benzene_t.json",
-        optimization_keywords=exess.OptimizationKeywords(),
+        optimization_keywords=exess_geo_opt.OptimizationKeywords(),
         method="RestrictedRIMP2",
         basis="cc-pVDZ",
         aux_basis="cc-pVDZ-RIFIT",
