+++ tests/test_exess_qmmm.py
@@ -1,14 +1,14 @@
 import sys
 from pathlib import Path
 
-from rush import exess
 from rush.client import RunOpts, save_object, set_opts
+from rush.exess_qmmm import exess_qmmm as run_exess_qmmm
 
 
 def test_exess_qmmm():
     set_opts(workspace_dir=Path.cwd() / "test-runs")
     data_dir = Path.cwd() / "tests" / "data"
-    res = exess.qmmm(
+    res = run_exess_qmmm(
         n_timesteps=500,
         temperature_kelvin=300.0,
         topology_path=data_dir / "6a5j_t.json",
@@ -16,7 +16,7 @@ def test_exess_qmmm():
         method="RestrictedHF",
         ksdft_keywords=None,
         # TODO: make this work (currently having convergence issues)
-        # restraints=exess.Restraints(free_fragments=[6]),
+        # restraints=exess_qmmm.Restraints(free_fragments=[6]),
         qm_fragments=[6],
         run_opts=RunOpts(
             name="Rush-Py Test EXESS QMMM 01: QM+MM",
